requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
streamlit==1.32.0
sqlalchemy==2.0.28
pymysql==1.1.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.31.0
bcrypt==4.1.2
